+++ NEWS.md
@@ -1,3 +1,9 @@
+# 5.0.0
+
+- removed the databases radon and radon-b and replaced it with the new, merged database rado.nb
+  - deprecated `get_radon` and `get_radonb`
+  - added a new function `get_rado.nb`
+
 # 4.1.0
 
 - added getter function for neonetatl database: `get_neonetatl`